[Refactor] Refactor tablet reshard structure with helper/splitter/merger modules
